+++ week_4/day_25/intcode.h
@@ -24,10 +24,10 @@ private:
    void read_opcode();
    void execute_instruction();
    std::vector<int> mode;
-   long instruction;
+   long long instruction;
    bool running;
    int  pos_in_memory;
-   long rb;
+   long long rb;
 
 
 public:
